Fixed separation edges appeared ugly
The previous approach to deal with separations was to add the reverse edges, but ignore those when computing the location of the nodes and edges, and adding edges in the opposite direction, which are hidden.

Now, instead, we use the dir="back" which is both simpler (no need for two edges per separation) and now the edges are much better (in the sense that they are much shorter, and easier to trace).